+++ tests/TestCase.php
@@ -26,6 +26,7 @@ use Nuwave\Lighthouse\Testing\MocksResolvers;
 use Nuwave\Lighthouse\Testing\TestingServiceProvider;
 use Nuwave\Lighthouse\Testing\UsesTestSchema;
 use Nuwave\Lighthouse\Validation\ValidationServiceProvider;
+use Nuwave\Lighthouse\Void\VoidServiceProvider;
 use Orchestra\Testbench\TestCase as BaseTestCase;
 use Symfony\Component\Console\Tester\CommandTester;
 use Tests\Utils\Policies\AuthServiceProvider;
@@ -83,6 +84,7 @@ GRAPHQL;
             SoftDeletesServiceProvider::class,
             TestingServiceProvider::class,
             ValidationServiceProvider::class,
+            VoidServiceProvider::class,
         ];
     }